feat: mark Avatar and AvatarGroup as stable
