--- .storybook/preview.ts
@@ -1,15 +1,0 @@
-import "../lib/css/global.css";
-
-/** @type { import('@storybook/react').Preview } */
-const preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
-
-export default preview;
